aula 18 insert de dados parte 1

+++ 03sqla_sync/models/sabor.py
@@ -7,7 +7,7 @@ from models.model_base import ModelBase
 class Sabor(ModelBase):
     __tablename__ : str = 'sabores'
 
-    id: int = sa.Column(sa.BigInteger, primary_key=True, autoincrement=True)
+    id: int = sa.Column(sa.Integer, primary_key=True, autoincrement=True)
     data_criacao: datetime = sa.Column(sa.DateTime, default=datetime.now, index=True)
 
     nome: str = sa.Column(sa.String(45), unique=True, nullable=False)
